requirements/base.txt
django ~= 4.0.1
djangorestframework ~=3.13.1
djoser ~= 2.1.0
djangorestframework-simplejwt < 5.0.0, >=4.3.0
django-environ >= 0.8.1,<1.0.0
django-cors-headers ~= 3.11.0
whitenoise ~= 5.3.0